+++ src/payment/index.js
@@ -6,6 +6,7 @@ const health = require('grpc-js-health-check')
 const opentelemetry = require('@opentelemetry/api')
 
 const charge = require('./charge')
+const refund = require('./refund')
 const logger = require('./logger')
 
 async function chargeServiceHandler(call, callback) {
@@ -29,6 +30,24 @@ async function chargeServiceHandler(call, callback) {
   }
 }
 
+async function refundServiceHandler(call, callback) {
+  const span = opentelemetry.trace.getActiveSpan();
+
+  try {
+    logger.info({ request: call.request }, "Refund request received.")
+
+    const response = await refund.refund(call.request)
+    callback(null, response)
+
+  } catch (err) {
+    logger.warn({ err })
+
+    span?.recordException(err)
+    span?.setStatus({ code: opentelemetry.SpanStatusCode.ERROR, message: err.message })
+    callback(err)
+  }
+}
+
 async function closeGracefully(signal) {
   server.forceShutdown()
   process.kill(process.pid, signal)
@@ -41,7 +60,10 @@ server.addService(health.service, new health.Implementation({
   '': health.servingStatus.SERVING
 }))
 
-server.addService(otelDemoPackage.oteldemo.PaymentService.service, { charge: chargeServiceHandler })
+server.addService(otelDemoPackage.oteldemo.PaymentService.service, {
+  charge: chargeServiceHandler,
+  refund: refundServiceHandler,
+})
 
 
 let ip = "0.0.0.0";